chore: Add .fake file to .gitignore and update yarn.lock
This commit adds the .fake file to the .gitignore file and updates the yarn.lock file. The .fake file was added to the .gitignore to exclude it from version control. The yarn.lock file was updated to reflect changes in the dependencies.

Note: This is a suggestion based on the code changes and recent commits. Please review and modify as needed.
Signed-off-by: Nickolas-Antoine Brochu <nickolas.antoine.brochu@dfrwmg.com>